useful python script
added python script to write the pages that contain links to all the exercises. Also added a script to write the template files for each exercise. Need to finish the script to include numbers of chapters that are not multiples of three and to include only the pdfs, rather than the typ files